cleanup project structure
