+++ internal/server/metrics.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package internal
+package server
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
